% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-avail.R
\name{avails}
\alias{avails}
\alias{br_avail_methods}
\alias{br_avail_methods_use_exp}
\alias{br_avail_method_config}
\title{Package availability}
\usage{
br_avail_methods()

br_avail_methods_use_exp()

br_avail_method_config(method)
}
\arguments{
\item{method}{Method for model construction. See \code{\link[=br_avail_methods]{br_avail_methods()}}
for available options.}
}
\value{
A character vector representing the available methods or options.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Package resource, definitions ready for use.
}
\section{Functions}{
\itemize{
\item \code{br_avail_methods()}: Returns available modeling methods. This correlates
to \code{\link[=br_set_model]{br_set_model()}}.

\item \code{br_avail_methods_use_exp()}: Returns available modeling methods which
set \code{exponentiate=TRUE} at default by \strong{bregr}.

\item \code{br_avail_method_config()}: Returns model configs for specified method to
generate modeling templates.

}}
\seealso{
\link{pipeline} for building \code{breg} objects.
}
