% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{br_set_x}
\alias{br_set_x}
\title{Set independent variables for model construction}
\usage{
br_set_x(obj, x, x2 = NULL)
}
\arguments{
\item{obj}{An object of class \code{breg}.}

\item{x}{Character vector representing focal variables.}

\item{x2}{Character vector representing control variables, optional.}
}
\description{
Set independent variables for model construction
}
