% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{br_set_y}
\alias{br_set_y}
\title{Set independent variables for model construction}
\usage{
br_set_y(y)
}
\arguments{
\item{y}{Character vector representing dependent variables.}
}
\description{
Set independent variables for model construction
}
