% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-bregr-package.R
\docType{package}
\name{bregr-package}
\alias{bregr}
\alias{bregr-package}
\title{bregr: Easy and Efficient Batch Processing of Regression Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Easily processes batches of univariate or multivariate regression models. Returns results in a tidy format and generates visualization plots for straightforward interpretation (Wang, Shixiang, et al. (2025) \doi{10.1002/mdr2.70028}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/bregr}
  \item \url{https://wanglabcsu.github.io/bregr/}
  \item Report bugs at \url{https://github.com/WangLabCSU/bregr/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

Authors:
\itemize{
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})
  \item Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})
  \item Chenyang Shu \email{402221506@qq.com}
}

}
\keyword{internal}
