% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcox-package.R
\docType{package}
\name{ezcox-package}
\alias{ezcox}
\alias{ezcox-package}
\title{ezcox: Easy and Efficient Batch Processing of Cox and Regression Models}
\description{
Easily process batches of univariate or multivariate Cox and regression models. Returns results in a tidy format and generates visualization plots for straightforward interpretation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/ezcox}
  \item Report bugs at \url{https://github.com/WangLabCSU/ezcox/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

Authors:
\itemize{
  \item Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})
  \item Chenyang Shu \email{402221506@qq.com}
}

}
\keyword{internal}
