% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-polar.R
\name{polar_connect}
\alias{polar_connect}
\title{Connects dots}
\usage{
polar_connect(data, mapping, ...)
}
\arguments{
\item{data}{A \code{data.frame} contains connections of all events.}

\item{mapping}{Set of aesthetic mappings to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.
Set mappings for \code{x} and \code{xend} are required.}

\item{...}{Other arguments passing to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Check \code{\link[=polar_init]{polar_init()}} for examples.
}
\seealso{
Other risk_network:
\code{\link[=br_show_risk_network]{br_show_risk_network()}},
\code{\link[=polar_init]{polar_init()}}
}
\concept{risk_network}
