% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-diagnostics.R
\name{print.br_diagnostics}
\alias{print.br_diagnostics}
\title{Print method for general diagnostic results}
\usage{
\method{print}{br_diagnostics}(x, ...)
}
\arguments{
\item{x}{A \code{br_diagnostics} object returned by \code{\link[=br_diagnose]{br_diagnose()}}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for general diagnostic results
}
