% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-diagnostics.R
\name{print.br_ph_test}
\alias{print.br_ph_test}
\title{Print method for proportional hazards test results}
\usage{
\method{print}{br_ph_test}(x, ...)
}
\arguments{
\item{x}{A \code{br_ph_test} object returned by \code{\link[=br_test_ph]{br_test_ph()}}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for proportional hazards test results
}
