% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-class.R
\name{print.breg}
\alias{print.breg}
\title{Print method for breg object}
\usage{
## S7 method for class <bregr::breg>
print(x, ..., raw = FALSE)
}
\arguments{
\item{x}{An object of class \code{breg}.}

\item{...}{Additional arguments (currently not used).}

\item{raw}{Logical, whether to print raw S7 representation. Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
