% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-compare.R
\name{print.breg_comparison}
\alias{print.breg_comparison}
\title{Print method for breg_comparison object}
\usage{
\method{print}{breg_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{breg_comparison} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for breg_comparison object
}
