% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-accessors.R
\name{accessors}
\alias{accessors}
\alias{br_get_data}
\alias{br_get_y}
\alias{br_get_x}
\alias{br_get_n_x}
\alias{br_get_n_x2}
\alias{br_get_x2}
\alias{br_get_group_by}
\alias{br_get_config}
\alias{br_get_models}
\alias{br_get_model}
\alias{br_get_results}
\title{Data accessors}
\usage{
br_get_data(obj)

br_get_y(obj)

br_get_x(obj)

br_get_n_x(obj)

br_get_n_x2(obj)

br_get_x2(obj)

br_get_group_by(obj)

br_get_config(obj)

br_get_models(obj)

br_get_model(obj, idx)

br_get_results(obj, tidy = FALSE, ...)
}
\arguments{
\item{obj}{A \code{breg} object.}

\item{idx}{Index or names (focal variables) of the model(s) to return.}

\item{tidy}{If \code{TRUE} return tidy (compact) results, otherwise return comprehensive results.
The tidy results are obtained from \code{\link[broom:reexports]{broom::tidy()}} while comprehensive results are obtained from
\code{\link[broom.helpers:tidy_plus_plus]{broom.helpers::tidy_plus_plus()}}. The results can be configured when run with \code{\link[=br_run]{br_run()}}.}

\item{...}{Subset operations passing to \code{\link[dplyr:filter]{dplyr::filter()}} to filter results.}
}
\description{
Accessors for \code{breg} objects. Opposites to \verb{br_set_*()} functions.
Functions with additional arguments provide additional functionality.
}
\seealso{
\link{pipeline} for building \code{breg} objects.
}
