% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-avail.R
\name{avails}
\alias{avails}
\alias{br_avail_methods}
\alias{br_avail_methods_use_exp}
\title{Package availability}
\usage{
br_avail_methods()

br_avail_methods_use_exp()
}
\value{
A character vector representing the available methods or options.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Package resource, definitions ready for use.
}
\section{Functions}{
\itemize{
\item \code{br_avail_methods()}: Returns available modeling methods. This correlates
to \code{\link[=br_set_model]{br_set_model()}}.

\item \code{br_avail_methods_use_exp()}: Returns available modeling methods which
set \code{exponentiate=TRUE} at default by \strong{bregr}.

}}
\seealso{
\link{pipeline} for building \code{breg} objects.
}
