% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_fitted_line}
\alias{br_show_fitted_line}
\title{Show fitted regression line with \code{visreg} interface}
\usage{
br_show_fitted_line(breg, idx = 1, ...)
}
\arguments{
\item{breg}{An object of class \code{breg} with results.}

\item{idx}{Length-1. Index or name (focal variable) of the model.}

\item{...}{Arguments passing to \code{\link[visreg:visreg]{visreg::visreg()}} excepts \code{fit}.}
}
\description{
Provide an interface to visualize the model results with \href{https://github.com/larmarange/ggstats/}{\strong{visreg}} package, to show how a predictor variable x affects an outcome y.
Illustration for arguments and examples could be found at \href{https://pbreheny.github.io/visreg/reference/visreg.html}{\code{visreg} reference page}, or please check the doc for dynamic dots \code{...}.
}
\seealso{
Other br_show: 
\code{\link{br_show_fitted_line_2d}()},
\code{\link{br_show_forest}()},
\code{\link{br_show_forest_ggstats}()},
\code{\link{br_show_forest_ggstatsplot}()},
\code{\link{br_show_risk_network}()},
\code{\link{br_show_table}()}
}
\concept{br_show}
