% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_fitted_line_2d}
\alias{br_show_fitted_line_2d}
\title{Show fitted regression line with \code{visreg} interface}
\usage{
br_show_fitted_line_2d(breg, idx = 1, ...)
}
\arguments{
\item{breg}{An object of class \code{breg} with results.}

\item{idx}{Length-1. Index or name (focal variable) of the model.}

\item{...}{Arguments passing to \code{\link[visreg:visreg2d]{visreg::visreg2d()}} excepts \code{fit}.}
}
\description{
Similar to \code{\link[=br_show_fitted_line]{br_show_fitted_line()}}, but visualize how two variables interact to affect the response in regression models.
Illustration for arguments and examples could be found at \href{https://pbreheny.github.io/visreg/reference/visreg2d.html}{\code{visreg2d} reference page}, or please check the doc for dynamic dots \code{...}.
}
\seealso{
Other br_show: 
\code{\link{br_show_fitted_line}()},
\code{\link{br_show_forest}()},
\code{\link{br_show_forest_ggstats}()},
\code{\link{br_show_forest_ggstatsplot}()},
\code{\link{br_show_risk_network}()},
\code{\link{br_show_table}()}
}
\concept{br_show}
