% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_forest}
\alias{br_show_forest}
\title{Show forest}
\usage{
br_show_forest(
  breg,
  clean = TRUE,
  ...,
  subset = NULL,
  drop = NULL,
  tab_headers = NULL
)
}
\arguments{
\item{breg}{An object of class \code{breg} with results.}

\item{clean}{If \code{TRUE}, remove "Group" or "Focal" variable column when the values in the result table
are same (before performing \code{subset} and \code{drop}),
and reduce repeat values in column "Group", "Focal", and "Variable".}

\item{...}{Arguments passing to \code{\link[forestploter:forest]{forestploter::forest()}}.}

\item{subset}{A expression to filter the result table rows for visualization.}

\item{drop}{A integer index for dropping result table columns for compact visualization.}

\item{tab_headers}{A character vector to replace plotted table.}
}
\description{
Show forest plot with \strong{forestploter} for fullfill the functionality of \strong{bregr} results.
}
\seealso{
Other br_show: 
\code{\link{br_show_fitted_line}()},
\code{\link{br_show_fitted_line_2d}()},
\code{\link{br_show_forest_ggstats}()},
\code{\link{br_show_forest_ggstatsplot}()},
\code{\link{br_show_risk_network}()},
\code{\link{br_show_table}()}
}
\concept{br_show}
