% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_forest}
\alias{br_show_forest}
\title{Show forest}
\usage{
br_show_forest(breg, ...)
}
\arguments{
\item{breg}{An object of class \code{breg} with results.}
}
\description{
Show forest
}
\seealso{
Other br_show: 
\code{\link{br_show_fitted_line}()},
\code{\link{br_show_fitted_line_2d}()},
\code{\link{br_show_forest_ggstats}()},
\code{\link{br_show_forest_ggstatsplot}()},
\code{\link{br_show_table}()}
}
\concept{br_show}
