% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_forest_ggstats}
\alias{br_show_forest_ggstats}
\title{Show forest with \code{ggstats} interface}
\usage{
br_show_forest_ggstats(breg, ..., idx = NULL)
}
\arguments{
\item{breg}{An object of class \code{breg} with results.}

\item{...}{Arguments passing to \code{\link[ggstats:ggcoef_model]{ggstats::ggcoef_table()}} or \code{\link[ggstats:ggcoef_model]{ggstats::ggcoef_compare()}}.}

\item{idx}{Index or names (focal variables) of the model(s).}
}
\description{
Provide an interface to visualize the model results with \href{https://github.com/larmarange/ggstats/}{\strong{ggstats}} package.
Illustration for arguments and examples could be found at \href{https://larmarange.github.io/ggstats/reference/ggcoef_model.html}{\code{ggcoef_model} reference page}, or please check the doc for \code{...}.
}
\concept{br_show}
