% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_forest_ggstats}
\alias{br_show_forest_ggstats}
\title{Show a forest plot with \code{ggstats} interface}
\usage{
br_show_forest_ggstats(breg, idx = NULL, ...)
}
\arguments{
\item{breg}{A regression object with results (must pass \code{assert_breg_obj_with_results()}).}

\item{idx}{Index or names (focal variables) of the model(s).}

\item{...}{Arguments passing to \code{\link[ggstats:ggcoef_model]{ggstats::ggcoef_table()}} or \code{\link[ggstats:ggcoef_model]{ggstats::ggcoef_compare()}} excepts \code{model}.}
}
\description{
Provides an interface to visualize the model results with \href{https://github.com/larmarange/ggstats/}{\strong{ggstats}} package.
Illustration for arguments and examples could be found at \href{https://larmarange.github.io/ggstats/reference/ggcoef_model.html}{\code{ggcoef_model} reference page}, or please check the doc for dynamic dots \code{...}.
}
\seealso{
Other br_show: 
\code{\link{br_show_fitted_line}()},
\code{\link{br_show_fitted_line_2d}()},
\code{\link{br_show_forest}()},
\code{\link{br_show_forest_ggstatsplot}()},
\code{\link{br_show_risk_network}()},
\code{\link{br_show_table}()}
}
\concept{br_show}
