% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_forest_ggstatsplot}
\alias{br_show_forest_ggstatsplot}
\title{Show forest with \code{ggstatsplot} interface}
\usage{
br_show_forest_ggstatsplot(breg, idx = 1, ...)
}
\arguments{
\item{breg}{An object of class \code{breg} with results.}

\item{idx}{Length-1. Index or name (focal variable) of the model(s).}

\item{...}{Arguments passing to \code{\link[ggstatsplot:ggcoefstats]{ggstatsplot::ggcoefstats()}}.}
}
\description{
Provide an interface to visualize the model results with \href{https://github.com/IndrajeetPatil/ggstatsplot/}{\strong{ggstatsplot}} package.
Illustration for arguments and examples could be found at \href{https://indrajeetpatil.github.io/ggstatsplot/reference/ggcoefstats.html}{\code{ggcoefstats} reference page}, or please check the doc for dynamic dots \code{...}.
}
\seealso{
Other br_show: 
\code{\link{br_show_forest_ggstats}()}
}
\concept{br_show}
