% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-show.R
\name{br_show_table}
\alias{br_show_table}
\title{Show model tidy results in table format}
\usage{
br_show_table(
  breg,
  ...,
  args_table_format = list(),
  export = FALSE,
  args_table_export = list()
)
}
\arguments{
\item{breg}{A regression object with results (must pass \code{assert_breg_obj_with_results()}).}

\item{...}{Arguments passing to \code{\link[=br_get_results]{br_get_results()}} for subsetting table.}

\item{args_table_format}{A list of arguments passing to \code{\link[insight:format_table]{insight::format_table()}}.}

\item{export}{Logical. If \code{TRUE}, show table for export purpose, e.g., present the table in Markdown or HTML format.}

\item{args_table_export}{A list of arguments passing to \code{\link[insight:export_table]{insight::export_table()}}. Only works when \code{export} is \code{TRUE}.}
}
\description{
Show model tidy results in table format
}
\seealso{
Other br_show: 
\code{\link{br_show_fitted_line}()},
\code{\link{br_show_fitted_line_2d}()},
\code{\link{br_show_forest}()},
\code{\link{br_show_forest_ggstats}()},
\code{\link{br_show_forest_ggstatsplot}()},
\code{\link{br_show_risk_network}()}
}
\concept{br_show}
