% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{breg}
\alias{breg}
\title{New breg-class object}
\usage{
breg(
  y = NULL,
  x = NULL,
  x2 = NULL,
  config = list(),
  models = list(),
  data = NULL
)
}
\arguments{
\item{y}{Character vector representing dependent variables.}

\item{x}{Character vector representing focal variables.}

\item{x2}{Character vector representing control variables, optional.}

\item{config}{Configurations for model construction.}

\item{models}{List of model.}

\item{data}{Data frame containing the model result data.}
}
\description{
New breg-class object
}
\examples{
obj <- breg("y", letters[1:5], LETTERS[1:5])
obj
print(obj, raw = TRUE)

}
