% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{breg}
\alias{breg}
\title{New breg-class object}
\usage{
breg(
  y = character(0),
  x = character(0),
  x2 = character(0),
  config = list(),
  models = list(),
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\arguments{
\item{y}{Character vector representing dependent variables.}

\item{x}{Character vector representing focal variables.}

\item{x2}{Character vector representing control variables, optional.}

\item{config}{List of model configurations.}

\item{data}{Data frame containing the model result data.}

\item{n_x}{Integer representing the number of focal variables.}

\item{n_x2}{Integer representing the number of control variables. description}

\item{raw}{Logical, whether to print raw S7 representation. Default is FALSE.}
}
\description{
New breg-class object
}
\examples{
obj = breg("y", "x1", c("x2", "x3"))

}
