% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{br_pipeline}
\alias{br_pipeline}
\alias{br_set_data}
\alias{br_set_y}
\alias{br_set_x}
\alias{br_set_model}
\alias{br_run}
\title{Modeling and analysis pipeline in bregr}
\usage{
br_pipeline(
  data,
  y,
  x,
  method,
  x2 = NULL,
  group_by = NULL,
  run_parallel = 1L,
  model_args = list(),
  run_args = list()
)

br_set_data(data)

br_set_y(obj, y)

br_set_x(obj, x, x2 = NULL)

br_set_model(obj, method, ...)

br_run(obj, ..., group_by = NULL, run_parallel = 1L)
}
\arguments{
\item{data}{A \code{data.frame} containing all necessary variables for analysis.
Column names should follow R's naming conventions.}

\item{y}{Character vector specifying dependent variables.}

\item{x}{Character vector specifying focal variables.}

\item{method}{Method for model construction.}

\item{x2}{Character vector specifying control variables (optional).}

\item{group_by}{A string specifying the group by column.}

\item{run_parallel}{Integer, indicating cores to run the task, default is \code{1}.}

\item{model_args}{A list of arguments passed to \code{br_set_model()}.}

\item{run_args}{A list of arguments passed to \code{br_run()}.}

\item{obj}{An object of class \code{breg}.}

\item{...}{Additional arguments for model construction in \code{br_set_model()}
or for \code{\link[parameters:model_parameters]{parameters::model_parameters()}} in \code{br_run()}.
Note: The default value for \code{exponentiate} is \code{FALSE} (coefficients are not exponentiated).
For coxph models, this is typically set to \code{TRUE}.}
}
\description{
\itemize{
\item \code{br_pipeline()}: All-in-one end to end pipeline to run the basic regression analysis in batch.
Which could be splitted into the following steps:
\item \code{br_set_data()}: Set data for model construction.
\item \code{br_set_y()}: Set independent variables for model construction.
\item \code{br_set_x()}: Set dependent variables for model construction.
\item \code{br_set_model()}: Set model configurations.
\item \code{br_run()}: Run the regression analysis in batch.
}
}
\examples{
data <- survival::lung
br_set_data(data) |>
  br_set_y(c("time", "status")) |>
  br_set_x(c("age", "sex"), "ph.ecog") |>
  br_set_model("coxph") |>
  br_run()

br_set_data(data) |>
  br_set_y(c("time", "status")) |>
  br_set_x(c("age", "sex"), "ph.ecog") |>
  br_set_model("coxph") |>
  br_run()

br_set_data(data) |>
  br_set_y(c("time", "status")) |>
  br_set_x(c("age * sex"), "pat.karno") |>
  br_set_model("coxph") |>
  br_run(exponentiate = TRUE, ci = 0.90)

br_set_data(data) |>
  br_set_y(c("time")) |>
  br_set_x(c("age", "sex"), "pat.karno") |>
  br_set_model('quasi(variance = "mu", link = "log")') |>
  br_run()

}
