% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{br_set_data}
\alias{br_set_data}
\alias{br_set_y}
\alias{br_set_x}
\alias{br_set_model}
\alias{br_run}
\title{Modeling and analysis pipeline in bregr}
\usage{
br_set_data(data)

br_set_y(obj, y)

br_set_x(obj, x, x2 = NULL)

br_set_model(obj, method, ...)

br_run(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{breg}.}

\item{y}{Character vector representing dependent variables.}

\item{x}{Character vector representing focal variables.}

\item{x2}{Character vector representing control variables, optional.}

\item{method}{Method for model construction.}

\item{...}{Additional arguments for model construction.}
}
\description{
\itemize{
\item \code{br_set_data()}: Set data for model construction.
\item \code{br_set_y()}: Set independent variables for model construction.
\item \code{br_set_x()}: Set dependent variables for model construction.
\item \code{br_set_model()}: Set model configurations.
\item \code{br_run()}: Run the regression analysis in batch.
}
}
\examples{
data <- survival::lung
br_set_data(data) |>
  br_set_y(c("time", "status")) |>
  br_set_x(c("age", "sex"), "ph.ecog") |>
  br_set_model("coxph") |>
  br_run()
}
