% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-polar.R
\name{polar_connect}
\alias{polar_connect}
\title{Connects dots}
\usage{
polar_connect(data, x1, x2, ...)
}
\arguments{
\item{data}{A \code{data.frame} contains connections of all events.}

\item{x1, x2}{Column names (\strong{without quote in \code{aes()}}) storing connected events.}

\item{...}{Arguments passing to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}},
expect \code{c(x, xend, y, yend)} these 4 mapping parameters.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Check \code{\link[=polar_init]{polar_init()}} for examples.
}
\seealso{
Other risk_network: 
\code{\link{br_show_risk_network}()},
\code{\link{polar_init}()}
}
\concept{risk_network}
